Fixed things up
Ok, when I mreged everything, there was a lot of errors. So I ask: are
you guys editing this in eclipse? Cause you should be.

There was a lot of confusion on the variable currency. In some methods,
the variable is a Currency, but in other methods its a String, which is
the name of the currency. Please fix this by Making proper use of the
variable. When referring to the name, call the variable currencyName.
Also, make the parameters either all String currencyName or Currency
currency.

I'll upload a version of the server so you guys can test your work